Implement SSR for Calendar (#3215)
