+++ testing/tests/DevExpress.serverSide/calendar.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/calendar.markup.tests.js");
