+++ testing/tests/DevExpress.ui.widgets.editors/calendarViews.tests.js
@@ -15,8 +15,6 @@ require("ui/calendar");
 
 var CALENDAR_EMPTY_CELL_CLASS = "dx-calendar-empty-cell",
     CALENDAR_CELL_CLASS = "dx-calendar-cell",
-    CALENDAR_TODAY_CLASS = "dx-calendar-today",
-    CALENDAR_OTHER_VIEW_CLASS = "dx-calendar-other-view",
     CALENDAR_SELECTED_DATE_CLASS = "dx-calendar-selected-date",
     CALENDAR_CONTOURED_DATE_CLASS = "dx-calendar-contoured-date",
 
@@ -27,15 +25,6 @@ var getShortDate = function(date) {
     return dateSerialization.serializeDate(date, dateUtils.getShortDateFormat());
 };
 
-var getTextsArray = function(elements) {
-    var result = [];
-
-    $.each(elements, function(_, element) {
-        result.push($(element).text());
-    });
-
-    return result;
-};
 
 var FakeView = BaseView.inherit({
     _isTodayCell: noop,
@@ -53,12 +42,6 @@ var FakeView = BaseView.inherit({
 
 
 QUnit.module("Basics");
-
-QUnit.test("all views must be derived from the base view class", function(assert) {
-    $.each(Views, function() {
-        assert.ok(new this($("<div>")) instanceof BaseView);
-    });
-});
 
 QUnit.test("onCellClick action should be fired on cell click", function(assert) {
     var $element = $("<div>").appendTo("body");
@@ -138,19 +121,6 @@ QUnit.module("MonthView", {
     }
 });
 
-QUnit.test("main table should be 6 by 7", function(assert) {
-    var table = this.$element.find("tbody");
-    assert.ok(table.length === 1, "table has been rendered");
-
-    var rows = table.find("tr");
-    assert.ok(rows.length === 6, "with 6 rows");
-
-    for(var i = 0; i < 6; ++i) {
-        var columns = $(rows[i]).find("td");
-        assert.ok(columns.length === 7, "of 7 columns");
-    }
-});
-
 QUnit.test("getNavigatorCaption must return a proper caption", function(assert) {
     assert.equal(this.view.getNavigatorCaption(), "October 2013", "caption is correct");
 });
@@ -158,62 +128,6 @@ QUnit.test("getNavigatorCaption must return a proper caption", function(assert) 
 QUnit.test("getNavigatorCaption must return a proper caption in RTL mode", function(assert) {
     this.view.option("rtl", true);
     assert.equal(this.view.getNavigatorCaption(), "October 2013", "caption is correct");
-});
-
-QUnit.test("day captions must be rendered in proper order", function(assert) {
-    var captions = this.$element.find("table").find("th");
-    assert.deepEqual(getTextsArray(captions), ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"], "day captions order is correct");
-});
-
-QUnit.test("day captions must be rendered in proper order in RTL mode", function(assert) {
-    this.reinit({
-        date: new Date(2013, 9, 16),
-        firstDayOfWeek: 1,
-        rtl: true
-    });
-
-    var captions = this.$element.find("table").find("th");
-    assert.deepEqual(getTextsArray(captions), ["Sun", "Sat", "Fri", "Thu", "Wed", "Tue", "Mon"], "day captions order is correct");
-});
-
-QUnit.test("dates must be rendered in proper positions", function(assert) {
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["30", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13",
-            "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27",
-            "28", "29", "30", "31", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]);
-});
-
-QUnit.test("dates must be rendered in proper positions in RTL mode", function(assert) {
-    this.reinit({
-        date: new Date(2013, 9, 16),
-        firstDayOfWeek: 1,
-        rtl: true
-    });
-
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["6", "5", "4", "3", "2", "1", "30", "13", "12", "11", "10", "9", "8", "7",
-            "20", "19", "18", "17", "16", "15", "14", "27", "26", "25", "24", "23", "22", "21",
-            "3", "2", "1", "31", "30", "29", "28", "10", "9", "8", "7", "6", "5", "4"]);
-});
-
-QUnit.test("dates must be rendered in proper positions when the first day of the month comes right before the first day of the week", function(assert) {
-    this.reinit({
-        date: new Date(2013, 8, 11),
-        firstDayOfWeek: 1
-    });
-
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["26", "27", "28", "29", "30", "31", "1", "2", "3", "4", "5", "6", "7", "8", "9",
-            "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23",
-            "24", "25", "26", "27", "28", "29", "30", "1", "2", "3", "4", "5", "6"]);
-});
-
-QUnit.test("non-current month dates must be decorated with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td").filter("." + CALENDAR_OTHER_VIEW_CLASS);
-    assert.deepEqual(getTextsArray(dateCells), ["30", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]);
 });
 
 QUnit.test("change value option must add a CSS class to a cell", function(assert) {
@@ -253,13 +167,6 @@ QUnit.test("changing contouredDate must remove CALENDAR_CONTOURED_DATE_CLASS cla
     assert.ok(!dateCell.hasClass(CALENDAR_CONTOURED_DATE_CLASS));
 });
 
-QUnit.test("today must be decorated with a css class", function(assert) {
-    this.reinit({});
-
-    var todayCell = this.$element.find("." + CALENDAR_TODAY_CLASS);
-    assert.equal(todayCell.length, 1);
-});
-
 QUnit.test("if option.disabled is set in a constructor, cells should not be clickable", function(assert) {
     assert.expect(0);
 
@@ -272,21 +179,6 @@ QUnit.test("if option.disabled is set in a constructor, cells should not be clic
     var date = this.$element.find("table").find("td")[0];
     pointerMock(date).click();
 });
-
-QUnit.test("value time component should not be compared in min and max options", function(assert) {
-    this.reinit({
-        value: new Date(2015, 2, 14, 12),
-        min: new Date(2015, 2, 14, 10)
-    });
-    assert.ok(!$(".dx-calendar-selected-date").hasClass("dx-calendar-empty-cell"), "current date is available");
-
-    this.reinit({
-        value: new Date(2015, 2, 14, 12),
-        max: new Date(2015, 2, 14, 13)
-    });
-    assert.ok(!$(".dx-calendar-selected-date").hasClass("dx-calendar-empty-cell"), "current date is available");
-});
-
 
 QUnit.module("YearView", {
     beforeEach: function() {
@@ -310,54 +202,8 @@ QUnit.module("YearView", {
     }
 });
 
-QUnit.test("main table for year view should be 4 by 3", function(assert) {
-    var table = this.$element.find("tbody");
-
-    assert.ok(table.length === 1, "table has been rendered");
-
-    var rows = table.find("tr");
-    assert.ok(rows.length === 3, "with 3 rows");
-
-    for(var i = 0; i < 3; ++i) {
-        var columns = $(rows[i]).find("td");
-        assert.ok(columns.length === 4, "of 4 columns");
-    }
-});
-
 QUnit.test("getNavigatorCaption must return a proper caption", function(assert) {
     assert.ok(this.view.getNavigatorCaption().toString() === "2013");
-});
-
-QUnit.test("month must be rendered in proper positions", function(assert) {
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["Jan", "Feb", "Mar", "Apr", "May",
-            "Jun", "Jul", "Aug", "Sep",
-            "Oct", "Nov", "Dec"]);
-});
-
-QUnit.test("month must be rendered in proper positions in RTL mode", function(assert) {
-    this.reinit({
-        date: new Date(2015, 2, 1),
-        rtl: true
-    });
-
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["Apr", "Mar", "Feb", "Jan",
-            "Aug", "Jul", "Jun", "May",
-            "Dec", "Nov", "Oct", "Sep"]);
-});
-
-QUnit.test("data-value after render for cells in year view", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        startMonth = 0;
-
-    $.each(dateCells, function(_, dateCell) {
-        var shortDate = getShortDate(new Date(2013, startMonth, 1));
-        assert.equal(shortDate, $(dateCell).data().value, "data-value has a current value");
-        startMonth++;
-    });
 });
 
 QUnit.test("change value option must add a CSS class to a cell", function(assert) {
@@ -388,14 +234,6 @@ QUnit.test("changing contouredDate must remove CALENDAR_CONTOURED_DATE_CLASS cla
     assert.ok(!dateCell.hasClass(CALENDAR_CONTOURED_DATE_CLASS));
 });
 
-QUnit.test("today must be decorated with a css class", function(assert) {
-    this.reinit({});
-
-    var todayCell = this.$element.find("." + CALENDAR_TODAY_CLASS);
-    assert.equal(todayCell.length, 1);
-});
-
-
 QUnit.module("DecadeView", {
     beforeEach: function() {
         fx.off = true;
@@ -419,48 +257,8 @@ QUnit.module("DecadeView", {
     }
 });
 
-QUnit.test("main table for decade view should be 4 by 3", function(assert) {
-    var table = this.$element.find("tbody");
-
-    assert.ok(table.length === 1, "table has been rendered");
-
-    var rows = table.find("tr");
-    assert.ok(rows.length === 3, "with 3 rows");
-
-    for(var i = 0; i < 3; ++i) {
-        var columns = $(rows[i]).find("td");
-        assert.ok(columns.length === 4, "of 4 columns");
-    }
-});
-
 QUnit.test("getNavigatorCaption must return a proper caption", function(assert) {
     assert.ok(this.view.getNavigatorCaption() === "2010-2019");
-});
-
-QUnit.test("years must be rendered in proper positions", function(assert) {
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["2009", "2010", "2011", "2012", "2013",
-            "2014", "2015", "2016", "2017",
-            "2018", "2019", "2020"]);
-});
-
-QUnit.test("years must be rendered in proper positions in RTL mode", function(assert) {
-    this.reinit({
-        date: new Date(2015, 2, 1),
-        rtl: true
-    });
-
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["2012", "2011", "2010", "2009", "2016",
-            "2015", "2014", "2013", "2020",
-            "2019", "2018", "2017"]);
-});
-
-QUnit.test("non-current decade dates must be decorated with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td").filter("." + CALENDAR_OTHER_VIEW_CLASS);
-    assert.deepEqual(getTextsArray(dateCells), ["2009", "2020"]);
 });
 
 QUnit.test("change value option must add a CSS class to a cell", function(assert) {
@@ -489,13 +287,6 @@ QUnit.test("changing contouredDate must remove CALENDAR_CONTOURED_DATE_CLASS cla
 
     this.view.option("contouredDate", newDate);
     assert.ok(!dateCell.hasClass(CALENDAR_CONTOURED_DATE_CLASS));
-});
-
-QUnit.test("today must be decorated with a css class", function(assert) {
-    this.reinit({});
-
-    var todayCell = this.$element.find("." + CALENDAR_TODAY_CLASS);
-    assert.equal(todayCell.length, 1);
 });
 
 QUnit.test("data-value after render for cells in decade view", function(assert) {
@@ -533,43 +324,8 @@ QUnit.module("CenturyView", {
     }
 });
 
-QUnit.test("main table for century view should be 4 by 3", function(assert) {
-    var table = this.$element.find("tbody");
-
-    assert.ok(table.length === 1, "table has been rendered");
-
-    var rows = table.find("tr");
-    assert.ok(rows.length === 3, "with 3 rows");
-
-    for(var i = 0; i < 3; ++i) {
-        var columns = $(rows[i]).find("td");
-        assert.ok(columns.length === 4, "of 4 columns");
-    }
-});
-
 QUnit.test("getNavigatorCaption must return a proper caption", function(assert) {
     assert.ok(this.view.getNavigatorCaption() === "2000-2099");
-});
-
-QUnit.test("decades must be rendered in proper positions", function(assert) {
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["1990 - 1999", "2000 - 2009", "2010 - 2019", "2020 - 2029", "2030 - 2039",
-            "2040 - 2049", "2050 - 2059", "2060 - 2069", "2070 - 2079",
-            "2080 - 2089", "2090 - 2099", "2100 - 2109"]);
-});
-
-QUnit.test("decades must be rendered in proper positions in RTL mode", function(assert) {
-    this.reinit({
-        date: new Date(2015, 2, 1),
-        rtl: true
-    });
-
-    var dateCells = this.$element.find("table").find("td");
-    assert.deepEqual(getTextsArray(dateCells),
-        ["2020 - 2029", "2010 - 2019", "2000 - 2009", "1990 - 1999", "2060 - 2069",
-            "2050 - 2059", "2040 - 2049", "2030 - 2039", "2100 - 2109",
-            "2090 - 2099", "2080 - 2089", "2070 - 2079"]);
 });
 
 QUnit.test("data-value after render for cells in century view", function(assert) {
@@ -581,19 +337,6 @@ QUnit.test("data-value after render for cells in century view", function(assert)
         assert.equal(shortDate, $(dateCell).data().value, "data-value has a current value");
         startYear += 10;
     });
-});
-
-QUnit.test("non-current century dates must be decorated with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td").filter("." + CALENDAR_OTHER_VIEW_CLASS);
-    assert.deepEqual(getTextsArray(dateCells), ["1990 - 1999", "2100 - 2109"]);
-});
-
-QUnit.test("change value option must add a CSS class to a cell", function(assert) {
-    var secondDate = new Date(2010, 1, 1),
-        secondDateCell = this.$element.find("table").find("td").eq(2);
-
-    this.view.option("value", secondDate);
-    assert.ok(secondDateCell.hasClass(CALENDAR_SELECTED_DATE_CLASS));
 });
 
 QUnit.test("changing contouredDate must add CALENDAR_CONTOURED_DATE_CLASS class to a cell", function(assert) {
@@ -616,14 +359,6 @@ QUnit.test("changing contouredDate must remove CALENDAR_CONTOURED_DATE_CLASS cla
     assert.ok(!dateCell.hasClass(CALENDAR_CONTOURED_DATE_CLASS));
 });
 
-QUnit.test("today must be decorated with a css class", function(assert) {
-    this.reinit({});
-
-    var todayCell = this.$element.find("." + CALENDAR_TODAY_CLASS);
-    assert.equal(todayCell.length, 1);
-});
-
-
 QUnit.module("MonthView min/max", {
     beforeEach: function() {
         fx.off = true;
@@ -643,13 +378,6 @@ QUnit.module("MonthView min/max", {
         this.$element.remove();
         fx.off = false;
     }
-});
-
-QUnit.test("monthView should not display dates earlier than min and later than max by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "31123426272829301234567891011");
 });
 
 QUnit.test("monthView should not allow to select dates earlier than min and later than max via pointer events", function(assert) {
@@ -700,12 +428,6 @@ QUnit.module("MonthView disabledDates", {
     }
 });
 
-QUnit.test("monthView should not display disabled dates by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "12341234");
-});
 
 QUnit.test("monthView should not allow to select disabled dates via pointer events", function(assert) {
     var disabledDays = [1, 2, 3, 4],
@@ -750,13 +472,6 @@ QUnit.module("MonthView disabledDates as array", {
     }
 });
 
-QUnit.test("monthView should not display disabled dates by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "1234");
-});
-
 QUnit.test("monthView should not allow to select disabled dates via pointer events", function(assert) {
     var disabledDays = [1, 2, 3, 4],
         dateCells = this.$element.find("table").find("td");
@@ -774,7 +489,6 @@ QUnit.test("monthView should not allow to navigate to a disabled date", function
     trigger(UP_ARROW_KEY_CODE);
     assert.deepEqual(view.option("contouredDate"), new Date(2010, 10, 5));
 });
-
 
 QUnit.module("YearView min/max", {
     beforeEach: function() {
@@ -794,17 +508,6 @@ QUnit.module("YearView min/max", {
         this.$element.remove();
         fx.off = false;
     }
-});
-
-QUnit.test("yearView should add empty_class for cells out of range ", function(assert) {
-    assert.equal(this.$element.find("." + CALENDAR_EMPTY_CELL_CLASS).length, 5, "correct empty cells count was rendered");
-});
-
-QUnit.test("yearView should not display dates earlier than min and later than max by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "AugSepOctNovDec");
 });
 
 QUnit.test("yearView should not allow to navigate to a date earlier than min and later than max via keyboard events", function(assert) {
@@ -845,17 +548,6 @@ QUnit.module("YearView disabledDates", {
     }
 });
 
-QUnit.test("yearView should add empty_class for disabled dates", function(assert) {
-    assert.equal(this.$element.find("." + CALENDAR_EMPTY_CELL_CLASS).length, 3, "correct empty cells count was rendered");
-});
-
-QUnit.test("yearView should not display disabled dates by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "JanFebMar");
-});
-
 QUnit.test("yearView should not allow to navigate to a disabled date via keyboard events", function(assert) {
     var $element = this.$element,
         view = this.view,
@@ -886,17 +578,6 @@ QUnit.module("DecadeView min/max", {
         this.$element.remove();
         fx.off = false;
     }
-});
-
-QUnit.test("decadeView should add empty_class for cells out of range ", function(assert) {
-    assert.equal(this.$element.find("." + CALENDAR_EMPTY_CELL_CLASS).length, 6, "correct empty cells count was rendered");
-});
-
-QUnit.test("decadeView should not display dates earlier than min and later than max by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "200920102011201220192020");
 });
 
 QUnit.test("decadeView should not allow to navigate to a date earlier than min and later than max via keyboard events", function(assert) {
@@ -937,17 +618,6 @@ QUnit.module("DecadeView disabledDates", {
     }
 });
 
-QUnit.test("decadeView should add empty_class for disabled dates", function(assert) {
-    assert.equal(this.$element.find("." + CALENDAR_EMPTY_CELL_CLASS).length, 4, "correct empty cells count was rendered");
-});
-
-QUnit.test("decadeView should not display disabled dates by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "2009201020112012");
-});
-
 QUnit.test("decadeView should not allow to navigate to a disabled date via keyboard events", function(assert) {
     var $element = this.$element,
         view = this.view,
@@ -977,17 +647,6 @@ QUnit.module("CenturyView min/max", {
         this.$element.remove();
         fx.off = false;
     }
-});
-
-QUnit.test("centuryView should add empty_class for cells out of range ", function(assert) {
-    assert.equal(this.$element.find("." + CALENDAR_EMPTY_CELL_CLASS).length, 4, "correct empty cells count was rendered");
-});
-
-QUnit.test("centuryView should not display dates earlier than min and later than max by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "1990 - 19992080 - 20892090 - 20992100 - 2109");
 });
 
 QUnit.test("centuryView should not allow to navigate to a date earlier than min and later than max via keyboard events", function(assert) {
@@ -1024,17 +683,6 @@ QUnit.module("CenturyView disabledDates", {
         this.$element.remove();
         fx.off = false;
     }
-});
-
-QUnit.test("centuryView should add empty_class for disabled dates", function(assert) {
-    assert.equal(this.$element.find("." + CALENDAR_EMPTY_CELL_CLASS).length, 2, "correct empty cells count was rendered");
-});
-
-QUnit.test("centuryView should not display disabled dates by decorating them with a CSS class", function(assert) {
-    var dateCells = this.$element.find("table").find("td"),
-        dateCellsText = dateCells.filter("." + CALENDAR_EMPTY_CELL_CLASS).text();
-
-    assert.equal(dateCellsText, "1990 - 19992000 - 2009");
 });
 
 QUnit.test("centuryView should not allow to navigate to a disabled date via keyboard events", function(assert) {
